Problem: Valid Anagram
Checking if 2 strings are anagrams by incrementing and decrementing a
vector of size 26 (english alphabet size). S char at i increments
the position of the letter in vector count while T char at i decrements
the position of the letter in vector. At the end if vector count is all
0 the strings are anagrams. This solution is O(n+m)=)(n) where n is size
of string S and m is size of string T. The space complexity is O(1)
because vector is always constant size of 26.