Problem: Two Sum
Looping through list and storing key/value pair (element/index).
Before the storing whe check if the (trget - current_element) is in
the map, if it is true that mean we have a pair of elements that
add up to the target and we return the pair of indices. If we loop
through the list and no pair is found return an empty list