Problem: Koko Eating Bananas
This function uses binary search to find the minimum eating speed k that
allows eating all piles within h hours. The search space for k is between 1
and the maximum number of items in any pile, which is found using
max_element(piles.begin(), piles.end()). In each iteration of the binary
search, the middle speed k is computed and the total hours required to eat
all piles at that speed is calculated by summing the hours for each pile
using ceil(static_cast<double>(pile) / k). If the total hours are less
than or equal to h, we try a smaller speed by adjusting the high bound.
If the total hours exceed h, we try a larger speed by adjusting the low
bound. The result is the smallest speed k that allows eating all
piles within the given time h. The time complexity is O(n log(m)) and
space complexity is O(1).