Problem: Kth Smallest Integer in BST
This code implements a solution to find the kth smallest element in a Binary
Search Tree (BST) using an in-order traversal approach. The inOrderTraversal
function recursively traverses the tree in an in-order manner, incrementing
a count variable each time a node is visited. When count equals k, it assigns
the current node's value to answer, which holds the kth smallest value. The
kthSmallest function initializes the count and answer variables, then calls
the inOrderTraversal function to find and return the kth smallest element in
the tree. This approach ensures an efficient traversal of the tree and finds
the desired element with a time complexity of O(n) and (h) space complexity
where n is the number of nodes.