Problem: Climbing Stairs
Classic 1D-DP problem that is solved bottom-up not using memoization
(top-down). This code solves the "climbing stairs" problem using
dynamic programming with constant space. It initializes two variables,
one (1 way to reach the 1st step) and two (2 ways to reach the 2nd step).
For n == 1 and n == 2, it returns the respective values directly. For n > 2,
the loop calculates the number of ways to reach each subsequent step by
summing the previous two steps (one + two). After each iteration, one and
two are updated to represent the last two steps. The final result holds the
number of ways to reach the nth step. The time complexity is O(n) as the
loop runs n-2 times, and the space complexity is O(1) since only a constant
amount of extra space is used.