Problem: Longest Turbulent Subarray
this solution is a type of kudanes algorithm. It works by calculating
the increasing and decreasing checksd (pos and neg). At the end of every
iteration the max of maxLen and max(pos,neg) is taken to see if maxLen
should be updated. if two elements are the same then pos and neg are
equal to 1 meaning the current turbulent sequence is over and current
turbulent sequence is checked again. The time complexity is O(n) while
space complexity is O(1)