Problem: Number of Sub-arrays of Size K and Average Greater than or Equal to Threshold
Using fixed sliding window pattern we have a vector that represents the  window
and total representing the total amout of windows the thier average is
>= threshold. we add elements to the window and check if the window size
exceeds k if it does we delete the first one the second if checks if the
window is size k and if the windows average is >= threshold, if it is
true total increments by 1. Ath the end after aff lements were processed
return total number of windows that satisfy the rule if they are >=
threshold. The time complexity is O(n) while space complexity is O(k).