Problem: Longest Substring Without Repeating Characters
Using sliding window approach this works by having an unordered_map with
char/int key value pair. While iterating through the string when a
duplicate is found the left pointer is calculated by taking the found
index of char + 1. The new position of the char is updated in thge map
and the result is recalculated by taking max(result, right - left + 1).
The +1 is needed because right - left is inclusive. THis process repeats
until there right == string size. The time complexity is O(n) while
space complexity is O(m) where m is size of uniqe chars in the string.