Problem: Longest Repeating Character Replacement
Implement a sliding window approach to find the longest
substring where at most k characters can be replaced to
make all characters the same. The algorithm keeps track
of character frequencies within the window and adjusts
the window size by moving the left pointer when the
number of characters to be replaced exceeds k. It ensures
that the result is the maximum length of such a valid substring.
THe time complexity is O(n) where the space complexity is O(m)