Problem: Middle of the Linked List
Given the head of a singly linked list, return the
middle node of the linked list. This is done by having a fast and slow
pointer. the fast pointer will move 2 times as fast as the slow pointer
meaning that when the fast pointer finishes the slow pointer will be at
the middle.