Problem: Maximum Twin Sum of a Linked List
The solution is quite trivial. First we find the middle of the linked
list, after that we reverse the remainder of the list so satisfy the
property of finding the twin node of i which is (n-1-i). then we iterate
from start of head and the start of the second half of the inked list
and compare ith node from first half and second half and store the max
of the sum of both. after that we return the max sum. The time and space
complexity is O(n).