Problem: Linked List Cycle II
We first find a node where fast and slow intersect and once the find is
ther fast is resedt to head node and increments one by one like slow
until fast and slow intersect again which will be the start of the cycle
(Floyd's Tortoise and Hare Algorithm). This approach is O(n) time and
O(1) space complexity.